feat: support --hash and --devtool
