+++ .gitignore
@@ -2,3 +2,4 @@
 .DS_Store
 tmp
 node_modules
+src/__fixtures__/**/dist
