+++ package.json
@@ -10,7 +10,7 @@
   "mai": "chencheng <sorrycc@gmail.com>",
   "license": "MIT",
   "bin": {
-    "antd-build": "./bin/antd-build"
+    "atool-build": "./bin/atool-build"
   },
   "scripts": {
     "publish": "git push origin master",
@@ -31,6 +31,8 @@
     "file-loader": "~0.8.4",
     "is-plain-object": "~2.0.1",
     "less-loader": "~2.2.0",
+    "map-json-webpack-plugin": "~1.0.0",
+    "rimraf": "~2.4.3",
     "shelljs": "~0.5.1",
     "url-loader": "~0.5.6",
     "webpack": "~1.12.2"
