+++ src/__fixtures__/build-normal/index.js
@@ -1,0 +1,2 @@
+console.log(1);
+
