+++ src/__fixtures__/build-normal/package.json
@@ -1,0 +1,5 @@
+{
+  "entry": {
+    "index": "./index.js"
+  }
+}
