+++ src/build.js
@@ -1,15 +1,21 @@
 import { join } from 'path';
-import shelljs from 'shelljs';
+import rimraf from 'rimraf';
 import webpack from 'webpack';
 import printResult from './printResult';
 import mergeCustomConfig from './mergeCustomConfig';
-import commonConfig from './webpack.common.config';
+import getWebpackCommonConfig from './getWebpackCommonConfig';
 
-export default function(args) {
-  const cwd = process.cwd();
-  const webpackConfig = mergeCustomConfig(commonConfig, cwd, 'production');
+const cwd = process.cwd();
 
-  // Config if not debug
+function getWebpackConfig(args) {
+  const webpackConfig = mergeCustomConfig(getWebpackCommonConfig(args), cwd, 'production');
+
+  // Config outputPath.
+  if (args.outputPath) {
+    webpackConfig.output.path = args.outputPath;
+  }
+
+  // Config if not debug.
   if (!args.debug) {
     webpackConfig.plugins = (webpackConfig.plugins || []).concat([
       new webpack.DefinePlugin({
@@ -23,13 +29,25 @@ export default function(args) {
     ]);
   }
 
-  // Config outputPath.
-  if (args.outputPath) {
-    webpackConfig.output.path = args.outputPath;
+  // Output map.json if hash.
+  if (args.hash) {
+    webpackConfig.output.filename = webpackConfig.output.chunkFilename = '[name]-[chunkhash].js';
+    webpackConfig.plugins = (webpackConfig.plugins).concat([
+      require('map-json-webpack-plugin')({
+        output: join(webpackConfig.output.path, 'map.json'),
+      }),
+    ]);
   }
 
+  return webpackConfig;
+}
+
+export default function(args) {
+  // Get config.
+  const webpackConfig = getWebpackConfig(args);
+
   // Clean output dir first.
-  shelljs.rm('-rf', join(cwd, webpackConfig.output.path));
+  rimraf.sync(webpackConfig.output.path);
 
   function doneHandler(err, stats) {
     printResult(stats);
